djangorest/requirements.txt
Django==3.0.4
psycopg2==2.8.4
djangorestframework==3.11.0
django-cors-headers==3.2.1
djangorestframework-simplejwt==4.4.0
kafka-python==2.0.1
